Add OSGi metadata to Kazuki, currently all packages are exported.
Note: I made the implementation bundle a fragment of the API bundle so they get merged together at runtime. This is because they each contribute different classes to the same packages, for example 'io.kazuki.v0.store.schema.SchemaStore' from the API bundle, and 'io.kazuki.v0.store.schema.SchemaStoreListener' from the implementation bundle. (This is known as a split package in OSGi, where the package content is split across multiple bundles.) The alternative solution was to use Require-Bundle, but that takes more effort and involves the use of import attributes.